Making use of standard contract HashPuzzleSha256.
